[Refactor] 인증 서버 gRPC 토큰 검증 시 Bearer 포함하여 받도록 수정
